ns3 transmet en tlv et ros reçoit en tlv

+++ ns-3/src/ros-ns3-main.cc
@@ -279,7 +279,7 @@ main (int argc, char *argv[])
 
   // NetAnim does not support creating nodes at run-time
   // We have to create nodes and then update them according to ROS
-  const uint32_t maxNodes = 5;
+  const uint8_t maxNodes = 5;
 
   NS_LOG_INFO("Initialisation des noeuds vehicules");
   initVehicules(maxNodes, ip_ROS);
