+++ ns-3/src/rosvehcomm/model/rosvehsync.h
@@ -27,7 +27,7 @@ namespace ns3 {
 
   struct SimInfo {
       std::string filename;
-      int nodeCount;
+      uint8_t nodeCount;
       Time duration;
   };
